Reorganize dirs

+++ build.sh
@@ -5,42 +5,20 @@ NAME=slurm-build
 TMP_DIST_DIR=./.tmp-dist
 DIST_DIR=./dist
 
-build () {
-        docker build --pull -t "$NAME" --file="$1" \
-                --build-arg DEBFULLNAME="${DEBFULLNAME}" \
-                --build-arg DEBEMAIL="${DEBEMAIL}" \
-                .
-        docker ps -q -a -f "name=$NAME" | xargs -r docker rm
-        docker create --name="$NAME" "$NAME"
-        rm -rf "$TMP_DIST_DIR"
-        docker cp "${NAME}:/dist" "$TMP_DIST_DIR"
-        docker rm "$NAME"
-        mkdir -p "$DIST_DIR"
-        cp "$TMP_DIST_DIR"/* "$DIST_DIR"
-        rm -rf "$TMP_DIST_DIR"
-}
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 dir"
+    exit 1
+fi
 
-build_bionic () {
-        build Dockerfile.bionic-ppa
-}
-
-build_centos7 () {
-        build Dockerfile.centos7
-}
-
-case "$1" in
-        bionic)
-                build_bionic
-                ;;
-        centos7)
-                build_centos7
-                ;;
-        all)
-                build_bionic
-                build_centos
-                ;;
-        *)
-                echo "Usage: $0 [bionic|centos7|all]"
-                exit 1
-                ;;
-esac
+docker build --pull -t "$NAME" \
+    --build-arg DEBFULLNAME="${DEBFULLNAME}" \
+    --build-arg DEBEMAIL="${DEBEMAIL}" \
+    "$1"
+docker ps -q -a -f "name=$NAME" | xargs -r docker rm
+docker create --name="$NAME" "$NAME"
+rm -rf "$TMP_DIST_DIR"
+docker cp "${NAME}:/dist" "$TMP_DIST_DIR"
+docker rm "$NAME"
+mkdir -p "$DIST_DIR"
+cp "$TMP_DIST_DIR"/* "$DIST_DIR"
+rm -rf "$TMP_DIST_DIR"
